[glSample2] Let s make it a rotate
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@7440 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux_plugins/ADM_videoFilters6_openGl/sample_fragment2/sampleGl.cpp
@@ -133,8 +133,8 @@ bool openGlSample::getNextFrame(uint32_t *fn,ADMImage *image)
     glProgramY->setUniformValue("myTextureU", 1); 
     glProgramY->setUniformValue("myTextureV", 2); 
     glProgramY->setUniformValue("myTextureY", 0); 
-    glProgramY->setUniformValue("myWidth", image->GetWidth(PLANAR_Y)); 
-    glProgramY->setUniformValue("myHeight", image->GetHeight(PLANAR_Y)); 
+    glProgramY->setUniformValue("myWidth", (GLfloat)image->GetWidth(PLANAR_Y)); 
+    glProgramY->setUniformValue("myHeight", (GLfloat)image->GetHeight(PLANAR_Y)); 
 
     uploadAllPlanes(image);
